+++ src/com/magento/idea/magento2plugin/indexes/JsIndex.java
@@ -9,10 +9,10 @@ import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.indexing.FileBasedIndex;
 import com.magento.idea.magento2plugin.reference.xml.PolyVariantReferenceBase;
 import com.magento.idea.magento2plugin.stubs.indexes.js.RequireJsIndex;
-import com.magento.idea.magento2plugin.util.magento.GetAllSubFilesOfVirtualFileUtil;
-import com.magento.idea.magento2plugin.util.magento.GetFilePathUtil;
-import com.magento.idea.magento2plugin.util.magento.GetModuleNameUtil;
-import com.magento.idea.magento2plugin.util.magento.GetModuleSourceFilesUtil;
+import com.magento.idea.magento2plugin.reference.provider.util.GetAllSubFilesOfVirtualFileUtil;
+import com.magento.idea.magento2plugin.reference.provider.util.GetFilePathUtil;
+import com.magento.idea.magento2plugin.reference.provider.util.GetModuleNameUtil;
+import com.magento.idea.magento2plugin.reference.provider.util.GetModuleSourceFilesUtil;
 import org.jetbrains.annotations.NotNull;
 import java.util.ArrayList;
 import java.util.Collection;
